Switched Mersenne Twist implemetations
Prompted by Visual Studio compilation issues, and helped along by the
vastly simpler implementation Makoto Matsumoto and Takuji Nishimura
provide, I've switched what PRNG library I'm using.

+++ random.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 
-uint32_t seed();
-uint32_t roll_die(uint32_t max);
-uint32_t rand_range(uint32_t min, uint32_t max);
-int32_t rand_delta(); /* returns -2, -1, 1 or 2, for map stuff */
+unsigned long int seed();
+unsigned int roll_die(unsigned int max);
+unsigned int rand_range(unsigned int min, unsigned int max);
+int rand_delta(); /* returns -2, -1, 1 or 2, for map stuff */
 
 #endif
